win32 project for freetype
