+++ Source/Core/FontFaceHandle.cpp
@@ -251,5 +251,52 @@ int FontFaceHandle::GetKerning(word lhs, word rhs) const
 }
 
 
+// Generates (or shares) a layer derived from a font effect.
+FontFaceLayer* FontFaceHandle::GenerateLayer(FontEffect* font_effect)
+{
+    // See if this effect has been instanced before, as part of a different configuration.
+    FontLayerMap::iterator i = layers.find(font_effect);
+    if (i != layers.end())
+        return i->second;
+
+    FontFaceLayer* layer = new FontFaceLayer();
+    layers[font_effect] = layer;
+
+    if (font_effect == NULL)
+    {
+        layer->Initialise(this);
+    }
+    else
+    {
+        // Determine which, if any, layer the new layer should copy its geometry and textures from.
+        FontFaceLayer* clone = NULL;
+        bool deep_clone = true;
+        String generation_key;
+
+        if (!font_effect->HasUniqueTexture())
+        {
+            clone = base_layer;
+            deep_clone = false;
+        }
+        else
+        {
+            generation_key = font_effect->GetName() + ";" + font_effect->GetGenerationKey();
+            FontLayerCache::iterator cache_iterator = layer_cache.find(generation_key);
+            if (cache_iterator != layer_cache.end())
+                clone = cache_iterator->second;
+        }
+
+        // Create a new layer.
+        layer->Initialise(this, font_effect, clone, deep_clone);
+
+        // Cache the layer in the layer cache if it generated its own textures (ie, didn't clone).
+        if (clone == NULL)
+            layer_cache[generation_key] = layer;
+    }
+
+    return layer;
+}
+
+
 }
 }
